[Cases][AttachmentV2] FTR test folder reorg (#277091)
## Summary

Prep for the xpack.cases.attachments.enabled flag flip. Today those
flag-sensitive tests live inside the shared `tests/common` suite that
both `config_basic` and `config_trial_common` load — so flipping the
default would break them.

This PR isolates all flag-sensitive attachment tests into two dedicated,
explicitly flag-pinned configs and makes the shared suite fully agnostic
to the flag. No behavior change and no new test coverage.

### What changed
Attachment ftr tests are organized into 2 folders:

- `tests/common/attachments_framework/` — the default suite (FF-ON,
unified registry → cases-attachments)
- `tests/common/attachments_framework_legacy/` — the FF-OFF suite
(endpoint, osquery, external_references, persistable_state), which
asserts the exact legacy cases-comments on-disk shape.

### Checklist

- [ ] Any text added follows [EUI's writing
guidelines](https://elastic.github.io/eui/#/guidelines/writing), uses
sentence case text and includes [i18n
support](https://github.com/elastic/kibana/blob/main/src/platform/packages/shared/kbn-i18n/README.md)
- [ ]
[Documentation](https://www.elastic.co/guide/en/kibana/master/development-documentation.html)
was added for features that require explanation or tutorials
- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [ ] If a plugin configuration key changed, check if it needs to be
allowlisted in the cloud and added to the [docker
list](https://github.com/elastic/kibana/blob/main/src/dev/build/tasks/os_packages/docker_generator/resources/base/bin/kibana-docker)
- [ ] This was checked for breaking HTTP API changes, and any breaking
changes have been approved by the breaking-change committee. The
`release_note:breaking` label should be applied in these situations.
- [ ] [Flaky Test
Runner](https://ci-stats.kibana.dev/trigger_flaky_test_runner/1) was
used on any tests changed
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)
- [x] Review the [backport
guidelines](https://docs.google.com/document/d/1VyN5k91e5OVumlc0Gb9RPa3h1ewuPE705nRtioPiTvY/edit?usp=sharing)
and apply applicable `backport:*` labels.

(cherry picked from commit a09cea8f0c1197de11977e11fbf3cf9ce1aa3376)
