+++ .buildkite/ftr-manifests/ftr_platform_stateful_configs.yml
@@ -190,7 +190,8 @@ enabled:
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_basic.ts
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial.ts
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial_common.ts
-  - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial_unified_attachments.ts
+  - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial_attachments.ts
+  - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial_attachments_legacy.ts
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_no_public_base_url.ts
   - x-pack/platform/test/cases_api_integration/spaces_only/config.ts
   - x-pack/platform/test/cases_api_integration/spaces_only/config_analytics_v2.ts
