+++ x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial.ts
@@ -7,6 +7,9 @@
 
 import { createTestConfig } from '../common/config';
 
+// Note: the attachments-framework API tests run under their own flag-pinned
+// configs (`config_trial_attachments.ts` =true, `config_trial_attachments_legacy.ts`
+// =false), not this config. The rest of the common suite runs under `config_trial_common.ts`.
 export default createTestConfig('security_and_spaces', {
   license: 'trial',
   ssl: true,
