+++ x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial_attachments.ts
@@ -7,10 +7,14 @@
 
 import { createTestConfig } from '../common/config';
 
+// Default attachments-framework suite: the unified registry writing to the
+// `cases-attachments` SO with the feature flag ON. Pinned `=true` so it is
+// unaffected by the plugin default. The byte-clean FF-OFF counterpart runs
+// under `config_trial_attachments_legacy.ts`.
 export default createTestConfig('security_and_spaces', {
   license: 'trial',
   ssl: true,
-  testFiles: [require.resolve('./tests/trial_unified_attachments')],
+  testFiles: [require.resolve('./tests/common/attachments_framework')],
   publicBaseUrl: true,
   kbnServerArgs: ['--xpack.cases.attachments.enabled=true'],
 });
