+++ x-pack/platform/test/cases_api_integration/security_and_spaces/tests/common/attachments_framework/dashboard.ts
@@ -6,29 +6,32 @@
  */
 
 import expect from '@kbn/expect';
-import { MAP_ATTACHMENT_TYPE, MAP_SO_TYPE } from '@kbn/cases-plugin/common/constants/attachments';
+import {
+  DASHBOARD_ATTACHMENT_TYPE,
+  DASHBOARD_SO_TYPE,
+} from '@kbn/cases-plugin/common/constants/attachments';
 import type { BulkCreateAttachmentsRequestV2 } from '@kbn/cases-plugin/common/types/api';
-import type { FtrProviderContext } from '../../../common/ftr_provider_context';
-import { postCaseReq, postCommentUserReq } from '../../../common/lib/mock';
+import type { FtrProviderContext } from '../../../../common/ftr_provider_context';
+import { postCaseReq, postCommentUserReq } from '../../../../common/lib/mock';
 import {
   bulkCreateAttachments,
   createCase,
   deleteAllCaseItems,
   findCaseUserActions,
-} from '../../../common/lib/api';
+} from '../../../../common/lib/api';
 
 export default ({ getService }: FtrProviderContext): void => {
   const supertest = getService('supertest');
   const es = getService('es');
 
-  describe('Map saved-object attachments', () => {
-    const mapPayload = {
-      type: MAP_ATTACHMENT_TYPE,
+  describe('Dashboard saved-object attachments', () => {
+    const dashboardPayload = {
+      type: DASHBOARD_ATTACHMENT_TYPE,
       owner: 'securitySolutionFixture',
-      attachmentId: 'map-1',
+      attachmentId: 'dashboard-1',
       metadata: {
-        title: 'My map',
-        soType: MAP_SO_TYPE,
+        title: 'My dashboard',
+        soType: DASHBOARD_SO_TYPE,
       },
     };
 
@@ -36,12 +39,12 @@ export default ({ getService }: FtrProviderContext): void => {
       await deleteAllCaseItems(es);
     });
 
-    it('returns a unified-shaped response when the batch contains a map attachment', async () => {
+    it('returns a unified-shaped response when the batch contains a dashboard attachment', async () => {
       const postedCase = await createCase(supertest, postCaseReq);
       const updatedCase = await bulkCreateAttachments({
         supertest,
         caseId: postedCase.id,
-        params: [mapPayload] as unknown as BulkCreateAttachmentsRequestV2,
+        params: [dashboardPayload] as unknown as BulkCreateAttachmentsRequestV2,
       });
 
       expect(updatedCase.comments?.length).to.be(1);
@@ -51,21 +54,21 @@ export default ({ getService }: FtrProviderContext): void => {
         metadata: { title: string; soType: string };
         owner: string;
       };
-      expect(attachment.type).to.eql(MAP_ATTACHMENT_TYPE);
-      expect(attachment.attachmentId).to.eql('map-1');
+      expect(attachment.type).to.eql(DASHBOARD_ATTACHMENT_TYPE);
+      expect(attachment.attachmentId).to.eql('dashboard-1');
       expect(attachment.metadata).to.eql({
-        title: 'My map',
-        soType: MAP_SO_TYPE,
+        title: 'My dashboard',
+        soType: DASHBOARD_SO_TYPE,
       });
       expect(attachment.owner).to.eql('securitySolutionFixture');
     });
 
-    it('records a comment user-action when a map attachment is created', async () => {
+    it('records a comment user-action when a dashboard attachment is created', async () => {
       const postedCase = await createCase(supertest, postCaseReq);
       await bulkCreateAttachments({
         supertest,
         caseId: postedCase.id,
-        params: [mapPayload] as unknown as BulkCreateAttachmentsRequestV2,
+        params: [dashboardPayload] as unknown as BulkCreateAttachmentsRequestV2,
       });
 
       const { userActions } = await findCaseUserActions({ supertest, caseID: postedCase.id });
@@ -75,20 +78,20 @@ export default ({ getService }: FtrProviderContext): void => {
       expect(attachmentUserAction.action).to.eql('create');
     });
 
-    it('returns a unified-shaped response for a mixed batch (user comment + map)', async () => {
+    it('returns a unified-shaped response for a mixed batch (user comment + dashboard)', async () => {
       const postedCase = await createCase(supertest, postCaseReq);
       const updatedCase = await bulkCreateAttachments({
         supertest,
         caseId: postedCase.id,
-        params: [postCommentUserReq, mapPayload] as unknown as BulkCreateAttachmentsRequestV2,
+        params: [postCommentUserReq, dashboardPayload] as unknown as BulkCreateAttachmentsRequestV2,
       });
 
       expect(updatedCase.comments?.length).to.be(2);
-      const mapAttachment = updatedCase.comments!.find(
-        (comment) => comment.type === MAP_ATTACHMENT_TYPE
+      const dashboardAttachment = updatedCase.comments!.find(
+        (comment) => comment.type === DASHBOARD_ATTACHMENT_TYPE
       ) as unknown as { attachmentId: string };
-      expect(mapAttachment).to.be.ok();
-      expect(mapAttachment.attachmentId).to.eql('map-1');
+      expect(dashboardAttachment).to.be.ok();
+      expect(dashboardAttachment.attachmentId).to.eql('dashboard-1');
 
       const userPartner = updatedCase.comments!.find(
         (comment) => comment.type === 'user' || comment.type === 'comment'
