+++ x-pack/platform/test/cases_api_integration/security_and_spaces/tests/common/attachments_framework_legacy/index.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { FtrProviderContext } from '../../../../common/ftr_provider_context';
+import { createSpacesAndUsers, deleteSpacesAndUsers } from '../../../../common/lib/authentication';
+
+export default ({ loadTestFile, getService }: FtrProviderContext): void => {
+  describe('cases attachments framework (feature flag OFF, byte-clean)', function () {
+    before(async () => {
+      await createSpacesAndUsers(getService);
+    });
+
+    after(async () => {
+      await deleteSpacesAndUsers(getService);
+    });
+
+    loadTestFile(require.resolve('./external_references'));
+    loadTestFile(require.resolve('./persistable_state'));
+    loadTestFile(require.resolve('./endpoint'));
+    loadTestFile(require.resolve('./osquery'));
+  });
+};
